Moved the now model structure to old package prooftemplate for proof templates.
