+++ backend/business-partner-agent/src/main/java/org/hyperledger/bpa/controller/ProofTemplateController.java
@@ -26,7 +26,7 @@ import io.micronaut.security.annotation.Secured;
 import io.micronaut.security.rules.SecurityRule;
 import io.micronaut.validation.Validated;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import org.hyperledger.bpa.controller.api.prooftemplates2.ProofTemplate;
+import org.hyperledger.bpa.controller.api.prooftemplates.ProofTemplate;
 import org.hyperledger.bpa.impl.ProofTemplateManager;
 import org.hyperledger.bpa.impl.verification.prooftemplates.ValidUUID;
 import org.hyperledger.bpa.model.BPAProofTemplate;
