+++ backend/business-partner-agent/src/main/java/org/hyperledger/bpa/impl/prooftemplates/NonRevocationApplicator.java
@@ -1,4 +1,4 @@
-package org.hyperledger.bpa.impl.prooftemplates.aries;
+package org.hyperledger.bpa.impl.prooftemplates;
 
 import io.micronaut.core.annotation.Nullable;
 import lombok.Builder;
