+++ backend/business-partner-agent/src/main/java/org/hyperledger/bpa/model/prooftemplate/BPAAttribute.java
@@ -16,16 +16,16 @@
  * limitations under the License.
  */
 
-package org.hyperledger.bpa.model.prooftemplate2;
+package org.hyperledger.bpa.model.prooftemplate;
 
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import io.micronaut.core.annotation.Introspected;
 import lombok.*;
-import org.hyperledger.bpa.controller.api.prooftemplates2.AttributeGroup;
-import org.hyperledger.bpa.impl.verification.prooftemplates.ValidAttributeGroup;
+import org.hyperledger.bpa.controller.api.prooftemplates.Attribute;
+import org.hyperledger.bpa.impl.verification.prooftemplates.ValidAttributeCondition;
 
 import javax.validation.Valid;
-import javax.validation.constraints.NotNull;
+import javax.validation.constraints.NotEmpty;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -35,24 +35,27 @@ import java.util.stream.Collectors;
 @Builder
 @Introspected
 @JsonTypeInfo(use = JsonTypeInfo.Id.CLASS)
-// using a concrete class instead of a generic list does not unmarshal correctly
-// see https://github.com/micronaut-projects/micronaut-data/issues/1064
-public class BPAAttributeGroups {
-    @Singular
+public class BPAAttribute {
+    @NotEmpty
+    private String name;
     @Valid
-    @NotNull
-    @ValidAttributeGroup
-    List<BPAAttributeGroup> attributeGroups;
+    @Singular
+    @ValidAttributeCondition
+    private List<BPACondition> conditions;
 
-    public List<AttributeGroup> toRepresentation() {
-        return attributeGroups.stream()
-                .map(BPAAttributeGroup::toRepresentation)
-                .collect(Collectors.toList());
+    public Attribute toRepresentation() {
+        return new Attribute(
+                name,
+                conditions.stream()
+                        .map(BPACondition::toRepresentation)
+                        .collect(Collectors.toList()));
     }
 
-    public static BPAAttributeGroups fromRepresentation(List<AttributeGroup> proofTemplate) {
-        return new BPAAttributeGroups(proofTemplate.stream()
-                .map(BPAAttributeGroup::fromRepresentation)
-                .collect(Collectors.toList()));
+    public static BPAAttribute fromRepresentation(Attribute attribute) {
+        return new BPAAttribute(
+                attribute.getName(),
+                attribute.getConditions().stream()
+                        .map(BPACondition::fromRepresentation)
+                        .collect(Collectors.toList()));
     }
 }
