+++ backend/business-partner-agent/src/test/java/org/hyperledger/bpa/controller/ProofTemplateControllerTest.java
@@ -24,9 +24,9 @@ import io.micronaut.http.client.HttpClient;
 import io.micronaut.http.client.annotation.Client;
 import io.micronaut.test.annotation.MockBean;
 import io.micronaut.test.extensions.junit5.annotation.MicronautTest;
-import org.hyperledger.bpa.controller.api.prooftemplates2.Attribute;
-import org.hyperledger.bpa.controller.api.prooftemplates2.AttributeGroup;
-import org.hyperledger.bpa.controller.api.prooftemplates2.ProofTemplate;
+import org.hyperledger.bpa.controller.api.prooftemplates.Attribute;
+import org.hyperledger.bpa.controller.api.prooftemplates.AttributeGroup;
+import org.hyperledger.bpa.controller.api.prooftemplates.ProofTemplate;
 import org.hyperledger.bpa.impl.aries.config.SchemaService;
 import org.hyperledger.bpa.model.BPASchema;
 import org.junit.jupiter.api.Assertions;
