+++ backend/business-partner-agent/src/test/java/org/hyperledger/bpa/impl/prooftemplate/aries/ProofTemplateConversionTestBase.java
@@ -21,7 +21,7 @@ import io.micronaut.test.annotation.MockBean;
 import org.hyperledger.aries.api.present_proof.PresentProofRequest;
 import org.hyperledger.bpa.RunWithAries;
 import org.hyperledger.bpa.impl.aries.config.SchemaService;
-import org.hyperledger.bpa.impl.prooftemplates.aries.ProofTemplateConversion;
+import org.hyperledger.bpa.impl.prooftemplates.ProofTemplateConversion;
 import org.hyperledger.bpa.model.BPASchema;
 import org.hyperledger.bpa.model.Partner;
 import org.hyperledger.bpa.repository.PartnerRepository;
