+++ backend/business-partner-agent/src/test/java/org/hyperledger/bpa/model/BPAAttributeGroupTest.java
@@ -22,10 +22,10 @@ import io.micronaut.test.annotation.MockBean;
 import io.micronaut.test.extensions.junit5.annotation.MicronautTest;
 import io.micronaut.validation.validator.Validator;
 import org.hyperledger.bpa.impl.aries.config.SchemaService;
-import org.hyperledger.bpa.model.prooftemplate2.BPAAttribute;
-import org.hyperledger.bpa.model.prooftemplate2.BPAAttributeGroup;
-import org.hyperledger.bpa.model.prooftemplate2.BPACondition;
-import org.hyperledger.bpa.model.prooftemplate2.ValueOperators;
+import org.hyperledger.bpa.model.prooftemplate.BPAAttribute;
+import org.hyperledger.bpa.model.prooftemplate.BPAAttributeGroup;
+import org.hyperledger.bpa.model.prooftemplate.BPACondition;
+import org.hyperledger.bpa.model.prooftemplate.ValueOperators;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
